location type crud

+++ src/main/java/com/cs544/project/repository/LocationTypeRepository.java
@@ -3,5 +3,5 @@ package com.cs544.project.repository;
 import com.cs544.project.domain.LocationType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationTypeRepository extends JpaRepository<LocationType, Long> {
+public interface LocationTypeRepository extends JpaRepository<LocationType, Integer> {
 }
